Updated build.gradle file
