+++ prober/build.gradle
@@ -16,7 +16,6 @@ apply plugin: 'java'
 
 createUberJar('deployJar', 'prober', 'google.registry.monitoring.blackbox.Prober')
 
-
 dependencies {
   def deps = rootProject.dependencyMap
 
@@ -45,7 +44,6 @@ dependencies {
     testCompile deps['junit:junit']
     testCompile deps['org.mockito:mockito-core']
     testCompile project(':third_party')
-    //testCompile project(path: ':core', configuration: 'testRuntime')
 
   // Include auto-value in compile until nebula-lint understands
   // annotationProcessor
@@ -55,4 +53,3 @@ dependencies {
   testAnnotationProcessor deps['com.google.dagger:dagger-compiler']
 }
 
-}
